HW2/requirements.txt
aiogram==3.24.0
aiohttp==3.13.2
python-dotenv==1.2.1
sqlalchemy==2.0.45
asyncpg==0.31.0
altair==6.0.0
vega_datasets==0.9.0
pandas==2.3.3
requests==2.32.5
vl-convert-python==1.8.0
